refactor: :recycle: вынести низкоуровневые функции graph в internal
connect/close/query/ensureIndex переезжают в internal/connection.ts,
index.ts становится реэкспортом. Подготовка к появлению
updateGraph/withGraph в index.ts — новые модули будут импортировать
низкоуровневые функции из internal без циклических зависимостей.
Поведение и публичный API не меняются.
